fixed tests on skills components
