+++ clientApp/src/app/components/profile/profile.component.html
@@ -18,7 +18,7 @@ The second md-card/ng-show is for other users to look at other trainers profiles
     </mat-toolbar-row>
 
     <!--info section-->
-    <mat-card-content *ngIf="trainer != undefined && skills != undefined">
+    <mat-card-content >
       <div class="row">
         <div class="col-lg-2">
           <img layout="column" style="border-radius: 50%;" alt=""
